requirements.txt
flask
python-telegram-bot
google-generativeai
